Move initialize functions to trainer.setup
